v1.1 Rock Paper Scissors Game
The bugs with the win counts have been cleaned up. The loop wasn't including the random option or player option so it was testing the same option against the same option for each round the user chose to play. Everything is working now but I still want to clean up some of the language and add some time delay to the code to improve user experience.