v1.3 Rock Paper Scissors Game
Added some more context for users. Imported time library to add a few delays to the script to improve user experience. Everything is working as intended so I think this will be the final update (at least for now) for this project so I can pivot to other mini projects to test my skills